Fix rebase conflicts
