+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/Hotspots/DisablingRequestValidationTest.cs
@@ -18,8 +18,15 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using FluentAssertions;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarAnalyzer.Common;
+using SonarAnalyzer.Helpers;
 using SonarAnalyzer.UnitTest.MetadataReferences;
 using SonarAnalyzer.UnitTest.TestFramework;
 using CS = SonarAnalyzer.Rules.CSharp;
@@ -31,21 +38,97 @@ namespace SonarAnalyzer.UnitTest.Rules
     public class DisablingRequestValidationTest
     {
         private const string AspNetMvcVersion = "5.2.7";
+        private const string WebConfig = "Web.config";
+        // for the tests that don't test the XML logic, to avoid test failures caused by unexpected issues from Web.config files
+        private const string FolderWithoutWebConfig = @"TestCases\WebConfig\Empty\";
 
         [TestMethod]
         [TestCategory("Rule")]
-        [TestCategory("Hotspot")]
         public void DisablingRequestValidation_CS() =>
-            Verifier.VerifyAnalyzer(@"TestCases\Hotspots\DisablingRequestValidation.cs",
-                new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
+            Verifier.VerifyAnalyzer(@"TestCases\DisablingRequestValidation.cs",
+                new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, FolderWithoutWebConfig),
                 additionalReferences: NuGetMetadataReference.MicrosoftAspNetMvc(AspNetMvcVersion));
 
         [TestMethod]
         [TestCategory("Rule")]
-        [TestCategory("Hotspot")]
-        public void DisablingRequestValidation_VB() =>
-            Verifier.VerifyAnalyzer(@"TestCases\Hotspots\DisablingRequestValidation.vb",
-                new VB.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
+        public void DisablingRequestValidation_CS_Disabled() =>
+            Verifier.VerifyNoIssueReported(@"TestCases\DisablingRequestValidation.cs",
+                new CS.DisablingRequestValidation(AnalyzerConfiguration.Hotspot, @"TestCases\WebConfig"),
                 additionalReferences: NuGetMetadataReference.MicrosoftAspNetMvc(AspNetMvcVersion));
+
+        [DataTestMethod]
+        [DataRow(@"TestCases\WebConfig\S5753Values")]
+        [DataRow(@"TestCases\WebConfig\Foo")]
+        [DataRow(@"TestCases\WebConfig\Corrupt")]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_CS_WebConfig(string root) =>
+            DiagnosticVerifier.VerifyExternalFile(
+                SolutionBuilder.Create().AddProject(AnalyzerLanguage.CSharp).GetCompilation(),
+                new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, root),
+                File.ReadAllText(Path.Combine(root, WebConfig)));
+
+        [DataTestMethod]
+        [DataRow(@"TestCases\WebConfig\MultipleFiles", "SubFolder")]
+        [DataRow(@"TestCases\WebConfig\S5753EdgeValues", "3.9", "5.6")]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_CS_WebConfig_SubFolders(string rootDirectory, params string[] subFolders)
+        {
+            var compilation = SolutionBuilder.Create().AddProject(AnalyzerLanguage.CSharp).GetCompilation();
+            var languageVersion = compilation.LanguageVersionString();
+
+            var allDiagnostics = DiagnosticVerifier.GetDiagnostics(
+                compilation,
+                new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, rootDirectory),
+                CompilationErrorBehavior.Default).ToList();
+
+            allDiagnostics.Should().NotBeEmpty();
+            var rootWebConfig = Path.Combine(rootDirectory, WebConfig);
+            VerifyResults(rootWebConfig, allDiagnostics, languageVersion);
+            foreach (var subFolder in subFolders)
+            {
+                var subFolderWebConfig = Path.Combine(rootDirectory, subFolder, WebConfig);
+                VerifyResults(subFolderWebConfig, allDiagnostics, languageVersion);
+            }
+        }
+
+        [TestMethod]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_CS_WebConfig_LowerCase() =>
+            DiagnosticVerifier.VerifyExternalFile(
+                SolutionBuilder.Create().AddProject(AnalyzerLanguage.CSharp).GetCompilation(),
+                new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, @"TestCases\WebConfig\LowerCase\"),
+                File.ReadAllText(@"TestCases\WebConfig\LowerCase\web.config"));
+
+        [TestMethod]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_VB() =>
+            Verifier.VerifyAnalyzer(@"TestCases\DisablingRequestValidation.vb",
+                new VB.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, FolderWithoutWebConfig),
+                additionalReferences: NuGetMetadataReference.MicrosoftAspNetMvc(AspNetMvcVersion));
+
+        [TestMethod]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_VB_Disabled() =>
+            Verifier.VerifyNoIssueReported(@"TestCases\DisablingRequestValidation.vb",
+                new VB.DisablingRequestValidation(AnalyzerConfiguration.Hotspot, @"TestCases\WebConfig"),
+                additionalReferences: NuGetMetadataReference.MicrosoftAspNetMvc(AspNetMvcVersion));
+
+        [TestMethod]
+        [TestCategory("Rule")]
+        public void DisablingRequestValidation_VB_WebConfig() =>
+            DiagnosticVerifier.VerifyExternalFile(
+                SolutionBuilder.Create().AddProject(AnalyzerLanguage.VisualBasic).GetCompilation(),
+                new VB.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled, @"TestCases\WebConfig\S5753Values\"),
+                File.ReadAllText(Path.Combine(@"TestCases\WebConfig\S5753Values", WebConfig)));
+
+        // Verifies the results for the given web.config file path.
+        private static void VerifyResults(string webConfigPath, IList<Diagnostic> allDiagnostics, string languageVersion)
+        {
+            var actualIssues = allDiagnostics.Where(d => d.Location.GetLineSpan().Path.EndsWith(webConfigPath));
+            var expectedIssues = new IssueLocationCollector().GetExpectedIssueLocations(GetLines(webConfigPath)).ToList();
+            DiagnosticVerifier.CompareActualToExpected(languageVersion, actualIssues, expectedIssues, false);
+        }
+
+        private static IEnumerable<TextLine> GetLines(string path) => SourceText.From(File.ReadAllText(path)).Lines;
     }
 }
